Changes Bank API to use Long for account balance.

+++ lab2-mes/src/main/kotlin/pt/um/tf/lab2/lab2mes/Bank.kt
@@ -1,6 +1,6 @@
 package pt.um.tf.lab2.lab2mes
 
 interface Bank {
-    fun movement(mov : Int) : Boolean
-    fun balance() : Int
+    fun movement(mov : Long) : Boolean
+    fun balance() : Long
 }
