+++ lab2-srv/src/main/kotlin/pt/um/tf/lab2/lab2srv/Account.kt
@@ -3,13 +3,13 @@ package pt.um.tf.lab2.lab2srv
 import pt.um.tf.lab2.lab2mes.Bank
 
 class Account : Bank {
-    override fun balance(): Int {
+    override fun balance(): Long {
         return balance
     }
 
-    private var balance : Int = 0
+    private var balance : Long = 0
 
-    override fun movement(mov : Int): Boolean {
+    override fun movement(mov : Long): Boolean {
         var res = true
         if (mov > 0) {
             balance += mov
